# Pipeline/mscs621-cloud9-master/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Cloud9 MSCS621 Concordance
  description: A simple concordance API.
  contact:
    email: ronald.dartey1@marist.edu
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: http://localhost:8080/mscs721/concordance/1.0.0
  description: SwaggerHub API Auto Mocking
tags:
- name: analysis
  description: Operations available for text analysis
paths:
  /analyze:
    post:
      parameters:
        - in: query
          name: calculate
          schema:
            type: boolean
          description: Force or do not force input to be analyzed even if it has been analyzed already
        - in: query
          name: store
          schema:
            type: boolean
          description: Store or do not store analyzed data in database
      tags:
      - analysis
      summary: Calculate
      description: Post text to generate concordance
      operationId: get_concordance
      requestBody:
        description: Text to be analyzed
        content:
          text/plain:
            schema:
              type: string
            example: The brown fox jumped over the brown log.
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/analyze_result'
        "400":
          description: invalid input, object invalid
      x-openapi-router-controller: swagger_server.controllers.analysis_controller
  /locate:
    post:
      parameters:
        - in: query
          name: calculate
          schema:
            type: boolean
          description: Force or do not force input to be analyzed even if it has been analyzed already
        - in: query
          name: store
          schema:
            type: boolean
          description: Store or do not store analyzed data in database
      tags:
      - analysis
      summary: Calculate
      description: Post text to generate concordance
      operationId: get_locations
      requestBody:
        description: Text to be analyzed
        content:
          text/plain:
            schema:
              type: string
            example: The brown fox jumped over the brown log.
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/locate_result'
        "400":
          description: invalid input, object invalid
      x-openapi-router-controller: swagger_server.controllers.analysis_controller
components:
  schemas:
    analyze_result:
      required:
      - concordance
      - input
      type: object
      properties:
        concordance:
          type: array
          example:
          - token: brown
            count: 2
          - token: fox
            count: 1
          - token: jumped
            count: 1
          - token: log
            count: 1
          - token: over
            count: 1
          - token: the
            count: 2
          items:
            $ref: '#/components/schemas/analyze_result_concordance'
        input:
          type: string
          example: The brown fox jumped over the brown log.
      example:
        input: The brown fox jumped over the brown log.
        concordance:
        - token: brown
          count: 2
        - token: fox
          count: 1
        - token: jumped
          count: 1
        - token: log
          count: 1
        - token: over
          count: 1
        - token: the
          count: 2
    locate_result:
      required:
      - concordance
      - input
      type: object
      properties:
        concordance:
          type: array
          example:
          - token: brown
            locations: [1, 6]
          - token: fox
            locations: [2]
          - token: jumped
            locations: [3]
          - token: log
            locations: [7]
          - token: over
            locations: [4]
          - token: the
            locations: [0, 5]
          items:
            $ref: '#/components/schemas/locate_result_concordance'
        input:
          type: string
          example: The brown fox jumped over the brown log.
      example:
        input: The brown fox jumped over the brown log.
        concordance:
          - token: brown
            locations: [1, 6]
          - token: fox
            locations: [2]
          - token: jumped
            locations: [3]
          - token: log
            locations: [7]
          - token: over
            locations: [4]
          - token: the
            locations: [0, 5]
    analyze_result_concordance:
      type: object
      properties:
        token:
          type: string
        count:
          type: integer
    locate_result_concordance:
      type: object
      properties:
        token:
          type: string
        locations:
          type: array
          items:
            $ref: '#/components/schemas/concordance_indexes'
    concordance_indexes:
      type: integer